move navigation folder and whatsnew folder into layout folder - lots of import updates
