+++ frontend/src/components/layout/navigation/AppPicker.module.scss
@@ -1,4 +1,4 @@
-@import "../../styles/tokens.scss";
+@import "../../../styles/tokens.scss";
 @import "~@mozilla-protocol/core/protocol/css/includes/lib";
 
 .wrapper {
