+++ frontend/src/components/layout/navigation/MenuToggle.module.scss
@@ -1,4 +1,4 @@
-@import "../../styles/tokens.scss";
+@import "../../../styles/tokens.scss";
 
 .toggle {
   [class*="is-dark"] & {
