+++ frontend/src/components/layout/navigation/MobileNavigation.module.scss
@@ -1,12 +1,11 @@
-@import "../../styles/tokens.scss";
+@import "../../../styles/tokens.scss";
 @import "~@mozilla-protocol/core/protocol/css/includes/lib";
 
 .mobile-menu {
   width: 100%;
   position: absolute;
   left: 0;
-  // setting the z-index low to allow mobile
-  // menu to slide behind main navigation header
+  // z-index of 2 (so it overlaps the hero image)
   z-index: 2;
   top: $navHeight;
   background: $color-white;
