+++ frontend/src/components/layout/navigation/Navigation.module.scss
@@ -1,4 +1,4 @@
-@import "../../styles/tokens.scss";
+@import "../../../styles/tokens.scss";
 
 .site-nav {
   display: flex;
