+++ frontend/src/components/layout/navigation/Navigation.tsx
@@ -2,12 +2,12 @@ import Link from "next/link";
 import { useRouter } from "next/router";
 import { useLocalization } from "@fluent/react";
 import styles from "./Navigation.module.scss";
-import { useIsLoggedIn } from "../../hooks/session";
+import { useIsLoggedIn } from "../../../hooks/session";
 import { SignUpButton } from "./SignUpButton";
 import { SignInButton } from "./SignInButton";
-import { useProfiles } from "../../hooks/api/profile";
+import { useProfiles } from "../../../hooks/api/profile";
 import { UpgradeButton } from "./UpgradeButton";
-import { WhatsNewMenu } from "../layout/whatsnew/WhatsNewMenu";
+import { WhatsNewMenu } from "./whatsnew/WhatsNewMenu";
 import { UserMenu } from "./UserMenu";
 import { AppPicker } from "./AppPicker";
 import { MenuToggle } from "./MenuToggle";
@@ -31,6 +31,7 @@ export const Navigation = ({ ...props }) => {
       aria-expanded={mobileMenuState}
       onClick={() => handleToggle()}
     >
+      {/* passing toggle state to show correct icon */}
       <MenuToggle toggleState={mobileMenuState} />
     </a>
   );
