+++ frontend/src/components/layout/navigation/SignInButton.module.scss
@@ -1,4 +1,4 @@
-@import "../../styles/tokens.scss";
+@import "../../../styles/tokens.scss";
 
 .sign-in-button {
   // We're using CSS's `currentColor` here because the colour
