+++ frontend/src/components/layout/navigation/SignUpButton.module.scss
@@ -1,4 +1,4 @@
-@import "../../styles/tokens.scss";
+@import "../../../styles/tokens.scss";
 
 .sign-up-button {
   display: none;
