+++ frontend/src/components/layout/navigation/UpgradeButton.module.scss
@@ -1,4 +1,4 @@
-@import "../../styles/tokens.scss";
+@import "../../../styles/tokens.scss";
 
 .upgrade-button {
   appearance: none;
