+++ frontend/src/components/layout/navigation/whatsnew/WhatsNewMenu.module.scss
@@ -1,4 +1,4 @@
-@import "../../../styles/tokens.scss";
+@import "../../../../styles/tokens.scss";
 
 .trigger {
   flex-shrink: 0;
